create function add product by admin

+++ pages/api/products/[id].js
@@ -2,7 +2,8 @@ import dbConnect from '../../../util/mongo'
 import Product from '../../../models/Product'
 
 export default async function handler(req, res) {
-    const { method, query:{id} } = req
+    const { method, query:{id}, cookies } = req
+    const token = cookies.token
 
     dbConnect()
 
@@ -18,6 +19,9 @@ export default async function handler(req, res) {
     }
  
     if( method === "PUT" ) {
+        if(!token || token !== process.env.TOKEN){
+            return res.status(401).json("Not authenticated!")
+        }
 
         try {
             const product = await Product.findByIdAndUpdate( id, req.body, {new: true} )
@@ -29,6 +33,9 @@ export default async function handler(req, res) {
     }
 
     if( method === "DELETE" ) {
+        if(!token || token !== process.env.TOKEN){
+            return res.status(401).json("Not authenticated!")
+        }
 
         try {
             await Product.findByIdAndDelete(id);
